some organization of websocket server handlers
