+++ server/src/websocket_server/mod.rs
@@ -1,8 +1,5 @@
 use actix::prelude::*;
-use rand::{self, rngs::ThreadRng, Rng};
-use std::collections::{HashMap, HashSet};
-use std::sync::atomic::{AtomicUsize, Ordering};
-use std::sync::Arc;
+pub mod game_server;
 pub mod session;
 
 pub const MAIN_CHAT_ROOM: &str = "main";
@@ -14,7 +11,7 @@ pub enum MessageContent {
 
 #[derive(Message)]
 #[rtype(result = "()")]
-pub struct Message(pub MessageContent);
+pub struct AppMessage(pub MessageContent);
 
 #[derive(Message)]
 #[rtype(result = "()")]
@@ -25,7 +22,7 @@ pub struct BinaryMessage {
 #[derive(Message)]
 #[rtype(usize)]
 pub struct Connect {
-    pub session_address: Recipient<Message>,
+    pub session_address: Recipient<AppMessage>,
 }
 
 #[derive(Message)]
@@ -61,160 +58,4 @@ impl actix::Message for ListRooms {
 pub struct Join {
     pub sender_id: usize,
     pub room_name: String,
-}
-
-#[derive(Debug)]
-pub struct ChatServer {
-    sessions: HashMap<usize, Recipient<Message>>,
-    rooms: HashMap<String, HashSet<usize>>,
-    rng: ThreadRng,
-    visitor_count: Arc<AtomicUsize>,
-}
-
-impl ChatServer {
-    pub fn new(visitor_count: Arc<AtomicUsize>) -> ChatServer {
-        let mut rooms = HashMap::new();
-        rooms.insert(MAIN_CHAT_ROOM.to_owned(), HashSet::new());
-
-        ChatServer {
-            sessions: HashMap::new(),
-            rooms,
-            rng: rand::thread_rng(),
-            visitor_count,
-        }
-    }
-}
-
-impl ChatServer {
-    fn send_message(&self, room: &str, message: &str, skip_id: usize) {
-        if let Some(sessions) = self.rooms.get(room) {
-            for id in sessions {
-                if *id == skip_id {
-                    continue;
-                }
-
-                if let Some(session_address) = self.sessions.get(id) {
-                    session_address.do_send(Message(MessageContent::Str(message.to_owned())));
-                }
-            }
-        }
-    }
-
-    fn send_byte_message(&self, room: &str, message: &Vec<u8>, skip_id: usize) {
-        if let Some(sessions) = self.rooms.get(room) {
-            for id in sessions {
-                if *id == skip_id {
-                    continue;
-                }
-                if let Some(addr) = self.sessions.get(id) {
-                    addr.do_send(Message(MessageContent::Bytes(message.to_vec())));
-                }
-            }
-        }
-    }
-}
-
-impl Actor for ChatServer {
-    /// We are going to use simple Context, we just need ability to communicate
-    /// with other actors.
-    type Context = Context<Self>;
-}
-
-impl Handler<Connect> for ChatServer {
-    type Result = usize;
-    fn handle(&mut self, message: Connect, _: &mut Context<Self>) -> Self::Result {
-        let id = self.rng.gen::<usize>();
-        self.sessions.insert(id, message.session_address);
-        println!("Someone joined, assigned the id {}", id);
-
-        self.send_message(MAIN_CHAT_ROOM, "Someone joined", 0);
-
-        self.rooms
-            .entry(MAIN_CHAT_ROOM.to_owned())
-            .or_default()
-            .insert(id);
-
-        let count = self.visitor_count.fetch_add(1, Ordering::SeqCst);
-        self.send_message(MAIN_CHAT_ROOM, &format!("Total visitors {count}"), 0);
-
-        id
-    }
-}
-
-impl Handler<Disconnect> for ChatServer {
-    type Result = ();
-    fn handle(&mut self, message: Disconnect, _: &mut Context<Self>) {
-        println!("Someone disconnected");
-        let mut rooms: Vec<String> = Vec::new();
-        self.visitor_count.fetch_sub(1, Ordering::SeqCst);
-
-        if self.sessions.remove(&message.sender_id).is_some() {
-            for (room_name, sessions) in &mut self.rooms {
-                if sessions.remove(&message.sender_id) {
-                    rooms.push(room_name.to_owned());
-                }
-            }
-        }
-
-        for room in rooms {
-            self.send_message(&room, "Someone disconnected", 0);
-        }
-    }
-}
-
-impl Handler<ClientMessage> for ChatServer {
-    type Result = ();
-    fn handle(&mut self, message: ClientMessage, _: &mut Context<Self>) {
-        println!("message received: {}", message.content);
-        self.send_message(&message.room, message.content.as_str(), message.sender_id);
-    }
-}
-
-impl Handler<ClientBinaryMessage> for ChatServer {
-    type Result = ();
-    fn handle(&mut self, message: ClientBinaryMessage, _: &mut Context<Self>) {
-        println!("message received: {:?}", message.content);
-        self.send_byte_message(&message.room, &message.content.clone(), message.sender_id);
-    }
-}
-
-impl Handler<ListRooms> for ChatServer {
-    type Result = MessageResult<ListRooms>;
-    fn handle(&mut self, _: ListRooms, _: &mut Context<Self>) -> Self::Result {
-        let mut rooms = Vec::new();
-
-        for key in self.rooms.keys() {
-            rooms.push(key.to_owned())
-        }
-
-        MessageResult(rooms)
-    }
-}
-
-impl Handler<Join> for ChatServer {
-    type Result = ();
-    fn handle(&mut self, message: Join, _: &mut Context<Self>) {
-        let Join {
-            sender_id,
-            room_name,
-        } = message;
-        let mut rooms = Vec::new();
-
-        for (n, sessions) in &mut self.rooms {
-            if sessions.remove(&sender_id) {
-                rooms.push(n.to_owned());
-            }
-        }
-
-        for room in rooms {
-            self.send_message(&room, "Someone disconnected", 0);
-        }
-
-        self.rooms
-            .entry(room_name.clone())
-            .or_default()
-            .insert(sender_id);
-
-        self.send_message(&room_name, "Someone connected", sender_id);
-    }
 }
